+++ apps/app-launch/src/main.ts
@@ -1,5 +1,3 @@
-import { NACP } from '@tootallnate/nacp';
-
 const ctx = screen.getContext('2d');
 ctx.font = '24px system-ui';
 ctx.fillStyle = 'white';
@@ -9,24 +7,19 @@ ctx.textBaseline = 'middle';
 const gap = 48;
 let x = gap;
 let y = gap;
-for (const app of Switch.applications) {
-	const nacp = new NACP(app.nacp);
-	const icon = new Image();
-	icon.onload = () => {
-		if (x + icon.width >= 1280) {
+for (const { icon, name } of Switch.Application) {
+	if (!icon) continue;
+	const img = new Image();
+	img.onload = () => {
+		if (x + img.width >= 1280) {
 			x = gap;
-			y += icon.height + 40 + gap;
+			y += img.height + 40 + gap;
 		}
-		ctx.drawImage(icon, x, y);
-		ctx.fillText(
-			nacp.title,
-			x + icon.width / 2,
-			y + icon.height + 20,
-			icon.width
-		);
-		x += icon.width + gap;
+		ctx.drawImage(img, x, y);
+		ctx.fillText(name, x + img.width / 2, y + img.height + 20, img.width);
+		x += img.width + gap;
 	};
-	const url = URL.createObjectURL(new Blob([app.icon]));
-	icon.src = url;
+	const url = URL.createObjectURL(new Blob([icon]));
+	img.src = url;
 	//app.launch()
 }