+++ packages/nro/src/main.ts
@@ -10,25 +10,14 @@ import {
 } from 'node:fs';
 import { pathToFileURL, fileURLToPath } from 'node:url';
 import { relative } from 'node:path';
-import parseAuthor from 'parse-author';
 import { NACP } from '@tootallnate/nacp';
+import { patchNACP } from '@nx.js/patch-nacp';
 import * as NRO from '@tootallnate/nro';
 import * as RomFS from '@tootallnate/romfs';
-import type { PackageJson } from 'types-package-json';
+import terminalImage from 'terminal-image';
 
 const cwd = process.cwd();
-const packageRoot = new URL(`${pathToFileURL(cwd)}/`);
-
-// Read the `package.json` file to get the information that will
-// be embedded in the NACP section of the output `.nro` file
-const packageJsonUrl = new URL('package.json', packageRoot);
-const packageJsonStr = readFileSync(packageJsonUrl, 'utf8');
-const packageJson: PackageJson & { titleId?: string } =
-	JSON.parse(packageJsonStr);
-const { titleId, name, version, author: rawAuthor } = packageJson;
-const author =
-	typeof rawAuthor === 'string' ? parseAuthor(rawAuthor) : rawAuthor;
-
+const appRoot = new URL(`${pathToFileURL(cwd)}/`);
 const isSrcMode = existsSync(new URL('../tsconfig.json', import.meta.url));
 const nxjsNroUrl = new URL(
 	isSrcMode ? '../../../nxjs.nro' : '../dist/nxjs.nro',
@@ -40,43 +29,40 @@ const nxjsNro = await NRO.decode(nxjsNroBlob);
 
 // Icon
 let icon = nxjsNro.icon;
-console.log(chalk.bold('Loading Icon:'));
+console.log(chalk.bold('Icon:'));
 const iconName = 'icon.jpg';
 try {
-	const iconUrl = new URL('icon.jpg', packageRoot);
+	const iconUrl = new URL(iconName, appRoot);
 	const iconData = readFileSync(iconUrl);
 	icon = new Blob([iconData]);
-	console.log(
-		`  Using ${chalk.bold(`"${iconName}"`)} file (${bytes(
-			icon.size
-		).toLowerCase()})`
-	);
 } catch (err: any) {
 	if (err.code !== 'ENOENT') {
 		throw err;
 	}
 	console.log(
-		`  ⚠️ No ${chalk.bold(
+		`⚠️  No ${chalk.bold(
 			`"${iconName}"`
 		)} file found. Default nx.js icon will be used.`
 	);
 }
+console.log(
+	await terminalImage.buffer(Buffer.from(await icon!.arrayBuffer()), {
+		height: 12,
+	})
+);
 
 // NACP
 const nacp = new NACP(await nxjsNro.nacp!.arrayBuffer());
 console.log();
 console.log(chalk.bold('Setting metadata:'));
-if (titleId) nacp.id = titleId;
-if (name) nacp.title = name;
-if (version) nacp.version = version;
-if (author?.name) nacp.author = author.name;
+patchNACP(nacp, new URL('package.json', appRoot));
 console.log(`  ID: ${chalk.green(nacp.id.toString(16).padStart(16, '0'))}`);
 console.log(`  Title: ${chalk.green(nacp.title)}`);
 console.log(`  Version: ${chalk.green(nacp.version)}`);
 console.log(`  Author: ${chalk.green(nacp.author)}`);
 
 // RomFS
-const romfsDir = new URL('romfs/', packageRoot);
+const romfsDir = new URL('romfs/', appRoot);
 const romfsDirPath = fileURLToPath(romfsDir);
 const romfs = await RomFS.decode(nxjsNro.romfs!);
 console.log();
@@ -114,11 +100,7 @@ try {
 	if (err.code !== 'ENOENT') throw err;
 }
 
-// Store the app's NACP file into the RomFS, since I don't
-// know how to reliably get it at runtime using libnx APIs
-romfs['.nacp'] = new Blob([nacp.buffer]);
-
-const outputNroName = `${name}.nro`;
+const outputNroName = `${nacp.title}.nro`;
 
 if (!(romfs['main.js'] instanceof Blob)) {
 	console.log();
@@ -131,7 +113,9 @@ if (!(romfs['main.js'] instanceof Blob)) {
 	);
 	console.log(
 		chalk.yellow(
-			`The entrypoint file ${chalk.bold(`"${name}.js"`)} will need to`
+			`The entrypoint file ${chalk.bold(
+				`"${nacp.title}.js"`
+			)} will need to`
 		)
 	);
 	console.log(
@@ -150,7 +134,7 @@ const outputNro = await NRO.encode({
 	nacp: new Blob([new Uint8Array(nacp.buffer)]),
 	romfs: await RomFS.encode(romfs),
 });
-const outputNroUrl = new URL(outputNroName, packageRoot);
+const outputNroUrl = new URL(outputNroName, appRoot);
 writeFileSync(outputNroUrl, Buffer.from(await outputNro.arrayBuffer()));
 console.log(
 	chalk.green(