+++ packages/nro/CHANGELOG.md
@@ -1,4 +1,27 @@
-# nxjs-pack
+# @nx.js/nro
+
+## 0.0.36
+
+### Patch Changes
+
+- Ensure `runtime.js.map` file is uploaded in CI ([`e14ed3588db54eb7e0a906aae896fd90e4b3ae21`](https://github.com/TooTallNate/nx.js/commit/e14ed3588db54eb7e0a906aae896fd90e4b3ae21))
+
+## 0.0.35
+
+## 0.0.34
+
+## 0.0.33
+
+### Patch Changes
+
+- Rename `nxjs-pack` to `@nx.js/nro` ([`c9c3e15c6f9532c49211acd09886b0e468f93b42`](https://github.com/TooTallNate/nx.js/commit/c9c3e15c6f9532c49211acd09886b0e468f93b42))
+
+- Render app icon in the terminal ([`3ac5b4d01dc85584f3ae90bf5529d78851af775f`](https://github.com/TooTallNate/nx.js/commit/3ac5b4d01dc85584f3ae90bf5529d78851af775f))
+
+- Add `@nx.js/patch-nacp` package ([`a79ded8f5a2e242521c30cde71bdc381528b03d1`](https://github.com/TooTallNate/nx.js/commit/a79ded8f5a2e242521c30cde71bdc381528b03d1))
+
+- Updated dependencies [[`a79ded8f5a2e242521c30cde71bdc381528b03d1`](https://github.com/TooTallNate/nx.js/commit/a79ded8f5a2e242521c30cde71bdc381528b03d1)]:
+  - @nx.js/patch-nacp@0.0.1
 
 ## 0.0.32